test: updated the testcases
modified testcases based on the current modifications made in the code base.
